Revert "ElmyraService: Remove search button from settings activity"
This reverts commit f5ef372c571d12fc2775d8590b46d1e4f66efe96.

As confusing as the settings search hierarchy may be, it's still far
more convenient to show the search button on every screen than to force
the user to go all the way back to the settings home page just to
search.

We've reverted this change on the ROM side, so revert it in
ElmyraService as well to make the settings UI match.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>